Move GetTerminalBlockHash to powchain engine
